+++ test/Queue._runCleanupCycle.spec.js
@@ -1,21 +1,20 @@
 'use strict';
 
-var redis = require('redis');
-var Queue = require('../lib/Queue');
+var queueFactory = require('./queueFactory');
 var assert = require('chai').assert;
 var Promise = require('bluebird');
 var sinon = require('sinon');
 
 describe('Queue._runCleanupCycle', function () {
 
-  var client = redis.createClient({
-    host: process.env.REDIS_HOST
-  });
-  var queue = new Queue('test', client);
+  var queue;
   var clock = null;
 
-  beforeEach(function (done) {
-    return client.flushall(done);
+  before(function () {
+    return queueFactory.create()
+      .then(function (_queue) {
+        queue = _queue;
+      });
   });
 
   afterEach(function () {